+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.48.0] - 2021-02-25
+
+### Added
+
+- hmrcUserResearchBanner
+- StandardPhaseBanner, StandardBetaBanner, StandardAlphaBanner
+
+### Changed
+
+- hmrcStandardHeader to support adding a phase and user research banner
+
+### Compatible with
+
+- [hmrc/hmrc-frontend v1.26.1](https://github.com/hmrc/hmrc-frontend/releases/tag/v1.26.1)
+- [hmrc/play-frontend-govuk v0.63.0](https://github.com/hmrc/play-frontend-govuk/releases/tag/v0.63.0)
+- [alphagov/govuk-frontend v3.11.0](https://github.com/alphagov/govuk-frontend/releases/tag/v3.11.0)
+
 ## [0.47.0] - 2021-02-25
 
 ### Fixed